backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "fastapi[all,standard]>=0.121.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0",
    "stream-chat>=4.29.0",
    "uvicorn>=0.38.0",
    "requests>=2.31.0",
    "sqlmodel>=0.0.8",
]
